+++ api/v1beta1/frr_node_state_types.go
@@ -26,8 +26,6 @@ type FRRNodeStateSpec struct {
 
 // FRRNodeStateStatus defines the observed state of FRRNodeState.
 type FRRNodeStateStatus struct {
-	// DesiredConfig represents the desired FRR config, as the result of translating all of the `FRRConfiguration`s for the given node.
-	DesiredConfig string `json:"desiredConfig,omitempty"`
 	// RunningConfig represents the current FRR running config, which is the configuration the FRR instance is currently running with.
 	RunningConfig string `json:"runningConfig,omitempty"`
 	// LastConversionResult is the status of the last translation between the `FRRConfiguration`s resources and FRR's configuration, contains "success" or an error.