+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@
 - Support a --always-block parameter. The parameter accepts a list of comma separated cidrs to always block. This is useful to protect well known cidrs such as pods or clusterIPs. ([PR 88](https://github.com/metallb/frr-k8s/pull/88))
 - Support restarting the webhook pod when the rotator updates its cert secret.([PR 100](https://github.com/metallb/frr-k8s/pull/100)) 
 
+- Remove the DesiredConfig field from the status API ([PR 110](https://github.com/metallb/frr-k8s/pull/110))
+
 ### Bug fixes
 
 - FRRConfigurations from namespaces different than the one the daemon is deployed on were not validated with other resoureces. ([PR 91](https://github.com/metallb/frr-k8s/pull/91))